requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.3
jinja2==3.1.4
httpx==0.27.2
python-dotenv==1.0.1
cleanlab-tlm==1.1.22
sentence-transformers==3.0.1
transformers==4.44.0
torch>=2.2.0
scikit-learn==1.5.1
numpy>=1.26
nltk==3.9.1
